feat: Bump version to 0.5.0 and update FFI ABI
- Updated FFI version from 0.4.1 to 0.5.0.
- Changed `QttyQuantity.unit` and `QttyDerivedQuantity.{numerator,denominator}` from `UnitId` enums to raw `u32` to eliminate undefined behavior from invalid discriminants.
- Updated tests and helper functions to reflect the new raw `u32` unit representation.
- Added validation methods for unit IDs in `QttyQuantity` and `QttyDerivedQuantity`.
- Enhanced integration tests to handle invalid unit scenarios gracefully.
- Removed `scalar-decimal` feature from dependencies as it is no longer supported.

+++ qtty-ffi/src/lib.rs
@@ -187,7 +187,7 @@ mod tests {
         let status = unsafe { qtty_quantity_convert(src, UnitId::Kilometer as u32, &mut out) };
         assert_eq!(status, QttyStatus::Ok);
         assert!((out.value - 1.0).abs() < 1e-12);
-        assert_eq!(out.unit, UnitId::Kilometer);
+        assert_eq!(out.unit, UnitId::Kilometer as u32);
     }
 
     #[test]
@@ -197,7 +197,7 @@ mod tests {
         let status = unsafe { qtty_quantity_convert(src, UnitId::Hour as u32, &mut out) };
         assert_eq!(status, QttyStatus::Ok);
         assert!((out.value - 1.0).abs() < 1e-12);
-        assert_eq!(out.unit, UnitId::Hour);
+        assert_eq!(out.unit, UnitId::Hour as u32);
     }
 
     #[test]
@@ -208,7 +208,7 @@ mod tests {
         let status = unsafe { qtty_quantity_convert(src, UnitId::Radian as u32, &mut out) };
         assert_eq!(status, QttyStatus::Ok);
         assert!((out.value - PI).abs() < 1e-12);
-        assert_eq!(out.unit, UnitId::Radian);
+        assert_eq!(out.unit, UnitId::Radian as u32);
     }
 
     #[test]
